Copy ntru fixes from recent commits

+++ crypto_kem/ntruhps2048677/clean/crypto_sort.c
@@ -8,7 +8,7 @@
 #define int32_MINMAX(a,b) \
     do { \
         int32_t ab = (b) ^ (a); \
-        int32_t c = (b) - (a); \
+        int32_t c = (int32_t)((int64_t)(b) - (int64_t)(a)); \
         c ^= ab & (c ^ (b)); \
         c >>= 31; \
         c &= ab; \
